refactor: rewrite forms from class component to hooks
