+++ my-app/src/App.tsx
@@ -5,7 +5,7 @@ import Main from 'components/Main/Main';
 import Layout from 'components/Layout/Layout';
 import About from 'components/About/About';
 import NotFoundPage from 'components/NotFound/NotFoundPage';
-import FormContainer from 'components/Form/FormContainer';
+import FormContainer from 'components/Form/FormContainerHooks';
 import { routes } from 'helpers/constants/routes';
 import { searchCharacters, loadCharacters } from 'helpers/controllers/getCharacters';
